badi giroux - (working) averageSampling test - changed to GradeData to double and testing for output (6 - samples creates shifting scores)

+++ src/main/java/com/COMP3004CMS/cms/report/GradeData.java
@@ -6,10 +6,10 @@ import com.COMP3004CMS.cms.report.classRequirments.StudentInfo;
  * Class that contains the data to be analyzed
  */
 public class GradeData {
-    int grade;
+    double grade;
     StudentInfo stu;
 
-    public GradeData(int grade, StudentInfo stu) {
+    public GradeData(double grade, StudentInfo stu) {
         this.grade = grade;
         this.stu = stu;
     }
